release_final/RELEASE_PACK/dashboard/requirements.txt
streamlit
pandas
psycopg2-binary
redis
plotly
streamlit-autorefresh